Update index.html
I did alot for this, so to start I added a change in the phone number and work email. I added the Education, Skills and Volunteer Experience in <Strong> as well as changed what I have and removed the word I and added dates to my experience. Most importantly I added my experience and skills.